+++ backend/src/main/java/com/prograngers/backend/dto/response/auth/kakao/KakaoTokenResponse.java
@@ -1,4 +1,4 @@
-package com.prograngers.backend.dto.response.auth;
+package com.prograngers.backend.dto.response.auth.kakao;
 
 import lombok.Data;
