+++ backend/src/main/java/com/prograngers/backend/dto/response/auth/kakao/KakaoUserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.prograngers.backend.dto.response.auth;
+package com.prograngers.backend.dto.response.auth.kakao;
 
 import com.prograngers.backend.entity.Member;
 import lombok.Data;
